[LLVMGPU] Track shared parallel dimensions better in reduction config
This patch moves IterationDimTracker from Codegen/LLVMGPU/KernelDispatch.cpp to
Codegen/Utils and uses it in reduction configuration. This allows us to
map iteration space dimensions between different operations instead of
assuming everything to be in a certain order.
