+++ compiler/src/iree/compiler/Codegen/Utils/BUILD.bazel
@@ -24,6 +24,7 @@ iree_compiler_cc_library(
         "LinalgOpInfo.cpp",
         "LinkingUtils.cpp",
         "MarkerUtils.cpp",
+        "SliceUtils.cpp",
         "Utils.cpp",
     ],
     hdrs = [
@@ -34,6 +35,7 @@ iree_compiler_cc_library(
         "LinalgOpInfo.h",
         "LinkingUtils.h",
         "MarkerUtils.h",
+        "SliceUtils.h",
         "Utils.h",
     ],
     deps = [
@@ -60,11 +62,13 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:Analysis",
         "@llvm-project//mlir:ArithDialect",
         "@llvm-project//mlir:ArithUtils",
+        "@llvm-project//mlir:DestinationStyleOpInterface",
         "@llvm-project//mlir:DialectUtils",
         "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:FunctionInterfaces",
         "@llvm-project//mlir:GPUDialect",
         "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:IndexingMapOpInterface",
         "@llvm-project//mlir:LinalgDialect",
         "@llvm-project//mlir:LinalgInterfaces",
         "@llvm-project//mlir:LinalgTransforms",
